a new xml export/import system, utilizes a genericized form that allows for just passing nested python types (lists,dictionaries, etc ) to the exporter, and getting the same back from the importer.
construction in turn follows from the python data structure, and is pluggable
by registering against a handler against path into the data structure.

compatibility for the new impl is in two stages, on old xml formats
we use the old parser for a release cycle, its now depreceated.

for the new format, compatibility will be via a series of python transforms
applied the nested python data structure.